chore: move install directory resolution to core module
